test: move Cucumber runners to a parent package
