+++ src/test/java/org/w3c/epubcheck/DebugCucumberTest.java
@@ -1,4 +1,4 @@
-package org.w3c.epubcheck.test;
+package org.w3c.epubcheck;
 
 import org.junit.runner.RunWith;
