+++ src/test/java/org/w3c/epubcheck/RunCucumberTest.java
@@ -1,4 +1,4 @@
-package org.w3c.epubcheck.test;
+package org.w3c.epubcheck;
 
 import org.junit.runner.RunWith;
 
@@ -6,7 +6,9 @@ import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(features = "classpath:", strict = true)
+@CucumberOptions(
+    features = "classpath:",
+    strict = true)
 public class RunCucumberTest
 {
